feat: Implementations of Diff utility which works with binary files
Xdelta3 is a set of tools designed to compute changes between binary files. These changes (delta files) are similar to the output of the 'diff' program, in that they may be used to store and transmit only the changes between files. The 'delta files' that Xdelta3 manages are stored in RFC3284 (VCDIFF) format.

Log: Implementations of Diff utility which works with binary files

Issue: https://github.com/deepin-community/sig-deepin-sysdev-team/issues/249